fix2: numeric lib not compatible with u_int8

+++ Source/SBox.h
@@ -8,9 +8,9 @@
 class SBox 
 {
 private:
-    std::vector<uint8_t> box;
+    std::vector<unsigned int> box;
 
-    SBox(const std::vector<uint8_t>& _box) : box(_box) {}
+    SBox(const std::vector<unsigned int>& _box) : box(_box) {}
 
 public:
     SBox(const size_t size = 16) //TODO: provide seed?
@@ -24,22 +24,22 @@ public:
 
     SBox inverse() const
     {
-        std::vector<uint8_t> inverse(box.size());
-        for (uint8_t i = 0; i < box.size(); ++i)
+        std::vector<unsigned int> inverse(box.size());
+        for (unsigned int i = 0; i < box.size(); ++i)
         {
             inverse.at(box.at(i)) = i;
         }
         return SBox(inverse);
     }
 
-    uint8_t substitute(uint8_t input) const 
+    unsigned int substitute(unsigned int input) const
     {
         return box.at(input);
     }
 
-    std::vector<uint8_t> substitute_block(const std::vector<uint8_t>& block) const 
+    std::vector<unsigned int> substitute_block(const std::vector<unsigned int>& block) const
     {
-        std::vector<uint8_t> output(block.size());
+        std::vector<unsigned int> output(block.size());
         for (size_t i = 0; i < block.size(); ++i) 
         {
             output.at(i) = substitute(block.at(i));
